change artifacts the result of 'publish', instead of 'packageBin'
If I'm going to use this as the official builds instead of a local thing,
I will need the hashes and the other files maven wants. Doc and Src artifacts
shouldn't be different than a local production, but just having appveyor do
it reduces step count, which is generally a good thing.