Type safety using JsonValue on value objects
Includes changing parser/builder interaction, as that was required to
separate simple and complex value types. This also prevents JsonParser
and CsvParser from being a single-pass parser. CborParser is still single
pass, but only because its input is mutable, which may not be the best
thing.

The tests with one builder and one parser all work perfectly; trying to
put more than one 'generic' builder together isn't so hot.